Migrate realm_create tests
