+++ libparsec/crates/protocol/tests/realm.rs
@@ -10,159 +10,6 @@ use libparsec_tests_fixtures::prelude::*;
 use libparsec_types::prelude::*;
 
 #[parsec_test]
-fn serde_realm_create_req() {
-    // Generated from Python implementation (Parsec v2.6.0+dev)
-    // Content:
-    //   cmd: "realm_create"
-    //   role_certificate: hex!("666f6f626172")
-    let raw = hex!(
-        "82a3636d64ac7265616c6d5f637265617465b0726f6c655f6365727469666963617465c406"
-        "666f6f626172"
-    );
-
-    let req = authenticated_cmds::realm_create::Req {
-        role_certificate: b"foobar".as_ref().into(),
-    };
-
-    let expected = authenticated_cmds::AnyCmdReq::RealmCreate(req);
-
-    let data = authenticated_cmds::AnyCmdReq::load(&raw).unwrap();
-
-    assert_eq!(data, expected);
-
-    // Also test serialization round trip
-    let raw2 = if let authenticated_cmds::AnyCmdReq::RealmCreate(data) = data {
-        data.dump().unwrap()
-    } else {
-        unreachable!()
-    };
-
-    let data2 = authenticated_cmds::AnyCmdReq::load(&raw2).unwrap();
-
-    assert_eq!(data2, expected);
-}
-
-#[parsec_test]
-#[case::ok(
-    // Generated from Python implementation (Parsec v2.6.0+dev)
-    // Content:
-    //   status: "ok"
-    &hex!(
-        "81a6737461747573a26f6b"
-    )[..],
-    authenticated_cmds::realm_create::Rep::Ok
-)]
-#[case::invalid_certification(
-    // Generated from Python implementation (Parsec v2.6.0+dev)
-    // Content:
-    //   reason: "foobar"
-    //   status: "invalid_certification"
-    &hex!(
-        "82a6726561736f6ea6666f6f626172a6737461747573b5696e76616c69645f636572746966"
-        "69636174696f6e"
-    )[..],
-    authenticated_cmds::realm_create::Rep::InvalidCertification {
-        reason: Some("foobar".to_owned()),
-    }
-)]
-#[case::invalid_raw(
-    // Generated from Python implementation (Parsec v2.6.0+dev)
-    // Content:
-    //   reason: "foobar"
-    //   status: "invalid_raw"
-    &hex!(
-        "82a6726561736f6ea6666f6f626172a6737461747573ac696e76616c69645f64617461"
-    )[..],
-    authenticated_cmds::realm_create::Rep::InvalidData {
-        reason: Some("foobar".to_owned()),
-    }
-)]
-#[case::not_found(
-    // Generated from Python implementation (Parsec v2.6.0+dev)
-    // Content:
-    //   reason: "foobar"
-    //   status: "not_found"
-    &hex!(
-        "82a6726561736f6ea6666f6f626172a6737461747573a96e6f745f666f756e64"
-    )[..],
-    authenticated_cmds::realm_create::Rep::NotFound {
-        reason: Some("foobar".to_owned()),
-    }
-)]
-#[case::already_exists(
-    // Generated from Python implementation (Parsec v2.6.0+dev)
-    // Content:
-    //   status: "already_exists"
-    &hex!(
-        "81a6737461747573ae616c72656164795f657869737473"
-    )[..],
-    authenticated_cmds::realm_create::Rep::AlreadyExists
-)]
-#[should_panic(expected = "missing field `backend_timestamp`")]
-#[case::bad_timestamp_legacy(
-    // Generated from Python implementation (Parsec v2.11.1+dev)
-    // Content:
-    //   status: "bad_timestamp"
-    //
-    // Note that raw data does not contain:
-    //  - ballpark_client_early_offset
-    //  - ballpark_client_late_offset
-    //  - backend_timestamp
-    //  - client_timestamp
-    // This was valid behavior in api v2 but is no longer valid from v3 onwards.
-    // The corresponding expected values used here are therefore not important
-    // since loading raw data should fail.
-    //
-    &hex!("81a6737461747573ad6261645f74696d657374616d70")[..],
-    authenticated_cmds::realm_create::Rep::BadTimestamp {
-        reason: None,
-        ballpark_client_early_offset: 0.,
-        ballpark_client_late_offset: 0.,
-        backend_timestamp: "2000-1-2T01:00:00Z".parse().unwrap(),
-        client_timestamp: "2000-1-2T01:00:00Z".parse().unwrap(),
-    }
-)]
-#[case::bad_timestamp(
-    // Generated from Python implementation (Parsec v2.11.1+dev)
-    // Content:
-    //   backend_timestamp: ext(1, 946774800.0)
-    //   ballpark_client_early_offset: 50.0
-    //   ballpark_client_late_offset: 70.0
-    //   client_timestamp: ext(1, 946774800.0)
-    //   status: "bad_timestamp"
-    //
-    &hex!(
-        "85b16261636b656e645f74696d657374616d70d70141cc375188000000bc62616c6c706172"
-        "6b5f636c69656e745f6561726c795f6f6666736574cb4049000000000000bb62616c6c7061"
-        "726b5f636c69656e745f6c6174655f6f6666736574cb4051800000000000b0636c69656e74"
-        "5f74696d657374616d70d70141cc375188000000a6737461747573ad6261645f74696d6573"
-        "74616d70"
-    )[..],
-    authenticated_cmds::realm_create::Rep::BadTimestamp {
-        reason: None,
-        ballpark_client_early_offset: 50.,
-        ballpark_client_late_offset: 70.,
-        backend_timestamp: "2000-1-2T01:00:00Z".parse().unwrap(),
-        client_timestamp: "2000-1-2T01:00:00Z".parse().unwrap(),
-    }
-)]
-fn serde_realm_create_rep(
-    #[case] raw: &[u8],
-    #[case] expected: authenticated_cmds::realm_create::Rep,
-) {
-    let data = authenticated_cmds::realm_create::Rep::load(raw).unwrap();
-
-    assert_eq!(data, expected);
-
-    // Also test serialization round trip
-    let raw2 = data.dump().unwrap();
-
-    let data2 = authenticated_cmds::realm_create::Rep::load(&raw2).unwrap();
-
-    assert_eq!(data2, expected);
-}
-
-#[parsec_test]
 fn serde_realm_stats_req() {
     // Generated from Python implementation (Parsec v2.6.0+dev)
     // Content:
